rename and move rpc client to separate file
