+++ src/client/mod.rs
@@ -1,81 +1,21 @@
 // MIDEN CLIENT
 // ================================================================================================
 
-#[cfg(not(any(test, feature = "mock")))]
-use crate::errors::RpcApiError;
 use crate::{config::ClientConfig, errors::ClientError, store::Store};
-#[cfg(not(any(test, feature = "mock")))]
-use miden_node_proto::{
-    requests::{SubmitProvenTransactionRequest, SyncStateRequest},
-    responses::{SubmitProvenTransactionResponse, SyncStateResponse},
-};
-
 use miden_tx::TransactionExecutor;
+#[cfg(not(any(test, feature = "mock")))]
+use rpc_client::RpcClient;
 
 pub mod accounts;
 pub mod chain_data;
 pub mod notes;
+#[cfg(not(any(test, feature = "mock")))]
+pub mod rpc_client;
 pub mod sync_state;
 pub mod transactions;
 
 // CONSTANTS
 // ================================================================================================
-
-#[cfg(not(any(test, feature = "mock")))]
-struct LazyRpcClient(
-    Option<miden_node_proto::rpc::api_client::ApiClient<tonic::transport::Channel>>,
-    String,
-);
-
-#[cfg(not(any(test, feature = "mock")))]
-impl LazyRpcClient {
-    pub fn new(config_endpoint: String) -> LazyRpcClient {
-        LazyRpcClient(None, config_endpoint)
-    }
-
-    /// Executes the specified sync state request and returns the response.
-    pub async fn sync_state(
-        &mut self,
-        request: impl tonic::IntoRequest<SyncStateRequest>,
-    ) -> std::result::Result<tonic::Response<SyncStateResponse>, ClientError> {
-        let rpc_api = self.rpc_api().await?;
-        rpc_api
-            .sync_state(request)
-            .await
-            .map_err(|err| ClientError::RpcApiError(RpcApiError::RequestError(err)))
-    }
-
-    pub async fn submit_proven_transaction(
-        &mut self,
-        request: impl tonic::IntoRequest<SubmitProvenTransactionRequest>,
-    ) -> std::result::Result<tonic::Response<SubmitProvenTransactionResponse>, ClientError> {
-        let rpc_api = self.rpc_api().await?;
-        rpc_api
-            .submit_proven_transaction(request)
-            .await
-            .map_err(|err| ClientError::RpcApiError(RpcApiError::RequestError(err)))
-    }
-
-    /// Takes care of establishing the rpc connection if not connected yet and returns a reference
-    /// to the inner ApiClient
-    async fn rpc_api(
-        &mut self,
-    ) -> Result<
-        &mut miden_node_proto::rpc::api_client::ApiClient<tonic::transport::Channel>,
-        ClientError,
-    > {
-        use miden_node_proto::rpc::api_client::ApiClient;
-
-        if self.0.is_some() {
-            Ok(self.0.as_mut().unwrap())
-        } else {
-            let rpc_api = ApiClient::connect(self.1.clone())
-                .await
-                .map_err(|err| ClientError::RpcApiError(RpcApiError::ConnectionError(err)))?;
-            Ok(self.0.insert(rpc_api))
-        }
-    }
-}
 
 /// A light client for connecting to the Miden rollup network.
 ///
@@ -89,7 +29,7 @@ impl LazyRpcClient {
 pub struct Client {
     /// Local database containing information about the accounts managed by this client.
     store: Store,
-    rpc_api: LazyRpcClient,
+    rpc_api: RpcClient,
     tx_executor: TransactionExecutor<crate::store::data_store::SqliteDataStore>,
 }
 
@@ -107,7 +47,7 @@ impl Client {
 
         Ok(Self {
             store: Store::new((&config).into())?,
-            rpc_api: LazyRpcClient::new(config.node_endpoint.to_string()),
+            rpc_api: RpcClient::new(config.node_endpoint.to_string()),
             tx_executor: TransactionExecutor::new(SqliteDataStore::new(Store::new(
                 (&config).into(),
             )?)),
